Completed site w/ Stripe payment implementation

+++ src/app/api/menu-items/route.js
@@ -1,34 +1,39 @@
 import mongoose from "mongoose";
 import { MenuItem } from "@/app/models/MenuItems";
+import { isAdmin } from "../auth/[...nextauth]/route";
 
 export async function POST(req) {
   mongoose.connect(process.env.MONGO_URL);
   const data = await req.json();
-  
-  const menuItemDoc = await MenuItem.create(data);
-  return Response.json(menuItemDoc);
+
+  if (await isAdmin()) {
+    const menuItemDoc = await MenuItem.create(data);
+    return Response.json(menuItemDoc);
+  }
 }
 
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL);
-  const {_id, ...data} = await req.json();
-  
-  await MenuItem.findByIdAndUpdate(_id, data)
-  return Response.json(true)
-}
+  const { _id, ...data } = await req.json();
 
+  if (await isAdmin()) {
+    await MenuItem.findByIdAndUpdate(_id, data);
+  }
+  return Response.json(true);
+}
 
 export async function GET(req) {
   mongoose.connect(process.env.MONGO_URL);
-  return Response.json(
-    await MenuItem.find()
-  )
+  return Response.json(await MenuItem.find());
 }
 
 export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
-    const _id = url.searchParams.get("_id");
+  const _id = url.searchParams.get("_id");
+
+  if (await isAdmin()) {
     await MenuItem.deleteOne({ _id });
-    return Response.json(true);
+  }
+  return Response.json(true);
 }
